[DFC-551] - Rename one-login-language-toggle to frontend-language-toggle
