+++ src/components/common-templates/base-page.njk
@@ -34,8 +34,8 @@
   }) }}
   {% block backLink %}
       {% if showLanguageToggle %}
-        {% from "one-login-language-toggle/macro.njk" import oneloginLanguageSelect %}
-        {{ oneloginLanguageSelect({
+        {% from "frontend-language-toggle/macro.njk" import languageSelect %}
+        {{ languageSelect({
           ariaLabel: translate("govuk.languageToggle.ariaLabel"),
           class:'govuk-!-margin-bottom-1',
           activeLanguage: htmlLang,
