+++ src/components/frontend-language-toggle/language-select.yaml
