+++ src/components/frontend-language-toggle/macro.njk
@@ -1,3 +1,3 @@
-{% macro oneloginLanguageSelect(params) %}
+{% macro languageSelect(params) %}
   {%- include "./template.njk" -%}
 {% endmacro %}
