+++ src/components/frontend-language-toggle/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@govuk-one-login/one-login-language-toggle",
+  "name": "@govuk-one-login/frontend-language-toggle",
   "version": "2.0.0",
   "description": "Reusable language toggle for GDS One Login",
   "main": "",
@@ -13,11 +13,11 @@
     "url": "git+https://github.com/govuk-one-login/di-fec-ga4.git"
   },
   "keywords": [
-    "one-login",
+    "govuk-one-login",
     "gds",
     "language",
-    "onelogin",
-    "one login"
+    "govuk-onelogin",
+    "govuk one login"
   ],
   "author": "DI Frontend Capability Team",
   "license": "ISC",
